Refactor - Screen 정리 (#24)
