+++ src/App.tsx
@@ -1,20 +1,23 @@
-/**
- * Sample React Native App
- * https://github.com/facebook/react-native
- *
- * @format
- */
-
 import React from 'react';
-import {RootStack} from './screens/RootStack';
-import {ThemeProvider} from 'styled-components';
-import {light} from './style/theme';
+import {createNativeStackNavigator} from '@react-navigation/native-stack';
+import {MainTab} from './screens/MainTab';
+
+const Stack = createNativeStackNavigator();
 
 function App(): JSX.Element {
   return (
-    <ThemeProvider theme={light}>
-      <RootStack />
-    </ThemeProvider>
+    <Stack.Navigator screenOptions={{headerBackTitle: '닫기'}}>
+      <Stack.Screen
+        name="MainTab"
+        component={MainTab}
+        options={{headerShown: false}}
+      />
+      {/* <Stack.Screen
+        name="GoalList"
+        component={GoalList}
+        options={{title: '챌린지 리스트'}}
+      /> */}
+    </Stack.Navigator>
   );
 }
