--- src/screens/GoalList.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import {Text, View} from 'react-native';
-
-const GoalList = () => {
-  return (
-    <View>
-      <Text>GoalList</Text>
-    </View>
-  );
-};
-
-export default GoalList;
