--- src/screens/Login.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import {Text, View} from 'react-native';
-
-const Login = () => {
-  return (
-    <View>
-      <Text>로그인</Text>
-    </View>
-  );
-};
-
-export default Login;
