--- src/screens/RootStack.tsx
@@ -1,29 +1,0 @@
-import React from 'react';
-import {createNativeStackNavigator} from '@react-navigation/native-stack';
-import {MainTab} from './MainTab';
-import Login from './Login';
-import GoalList from './GoalList';
-
-const Stack = createNativeStackNavigator();
-
-export const RootStack = () => {
-  return (
-    <Stack.Navigator screenOptions={{headerBackTitle: '닫기'}}>
-      <Stack.Screen
-        name="MainTab"
-        component={MainTab}
-        options={{headerShown: false}}
-      />
-      <Stack.Screen
-        name="Login"
-        component={Login}
-        options={{title: '로그인'}}
-      />
-      <Stack.Screen
-        name="GoalList"
-        component={GoalList}
-        options={{title: '챌린지 리스트'}}
-      />
-    </Stack.Navigator>
-  );
-};
